client: some communication refactor
The 'main' script no longer talks to the receiver, instead they both
read from the configuration to know the socket to connect to.

Instead of relying on socket timeout and checking for shutdown (which could
introduce a latency of up to 50ms), the 'main' script tells the 'receiver'
to quit through the server. This is a weird roundtrip, but makes things
cleaner

The main client no longer waits if the server does not exist, which never
made much sense.